Add files via upload
Layout Design for an Website - 
Instructions:
Create one HTML file (e.g. first_page.html) that has:
Navigation buttons
A heading as an image: To do this, you will need to design a name of a website logo.
A colour scheme for the web page: This colour scheme must be defined in the separate style sheet and linked.
A heading and two paragraphs of text
An unordered list: The unordered list should look like the following when shown by a browser:
An unordered list can be specified with the tags
Contains a number of list items that can be specified

For the image, you can use any .jpg or .png file.

Create another HTML file that contains a heading and a couple of paragraphs.
You could name this new file another_page.html.
Place the second HTML file in the same folder as the first HTML file.
After you have created the new HTML web page, add a link to the first web page so that the browser will load another_page.html when you click Go to the other page in the first web page.
Use the title attribute which specifies a text that emerges when the mouse cursor is moved above a link. This kind of text is called a tool tip. Modify the link you created in the previous step so that when the mouse cursor is over the link, the tool tip reads, "This leads you to another page."
It is possible to use a picture (image) as a link. Modify your web page so that the picture that is on your web page will also serve as a link that leads to another web page.
Make sure all the images you have included in the web pages are in a separate subfolder named /images.
Write the steps to upload the website on a server and test that your web pages work as real Internet web pages.