Add files via upload
Block of JavaScript code where,

The background colour of the home page can keep changing after regular intervals.
The current system date and time should be displayed on the home page at a location visible to everyone.
Create a new web page. Write a JavaScript code to display the number of days left until your birthday as follows:
Define a variable with your birthday.
The script should calculate the difference; in other words, the number of days until your birthday.
If the difference is less than 30 days, it should display the message in red. If the difference is 2–6 months, the message should be in green. Otherwise, the message should be in blue.